+++ recipes/_common-attributes.rb
@@ -1,6 +1,5 @@
 # Alfresco dir root (used in _alfrescoproperties-attributes.rb and below)
 node.default['alfresco']['properties']['dir.root'] = "#{node['alfresco']['home']}/alf_data"
-node.default['alfresco']['properties']['dir.keystore'] = "#{node['alfresco']['properties']['dir.root']}/keystore/alfresco/keystore"
 
 # Solr Common attributes (used in _tomcat-attributes.rb)
 node.default['alfresco']['solr']['alfresco_models'] = "#{node['alfresco']['properties']['dir.root']}/newAlfrescoModels"
@@ -25,3 +24,22 @@ end
 
 # If enabled, Tomcat SSL Connector will use this redirectPort
 node.default['tomcat']['ssl_redirect_port'] = node['alfresco']['public_portssl']
+
+# Logrotate values; they will be used only if logrotate::global (or a wrapping recipe)
+# is part of the run_list
+node.default['logrotate']['global']['/var/log/haproxy/*.log'] = {
+  'daily'  => true,
+  'weekly'  => false,
+  'create' => '600 haproxy haproxy',
+  'postrotate'  => ['[ -f /var/run/syslogd.pid ] && kill -USR1 `cat /var/run/syslogd.pid`']
+}
+
+node.default['logrotate']['global']['/var/log/nginx/*.log'] = {
+  'daily'  => true,
+  'weekly'  => false,
+  'delaycompress'  => true,
+  'notifempty' => true,
+  'sharedscripts' => true,
+  'create' => '600 nginx nginx',
+  'postrotate' => ['[ -f /var/run/nginx.pid ] && kill -USR1 `cat /var/run/nginx.pid`']
+}